+++ src/components/users/avatar/UserAvatar.tsx
@@ -1,18 +1,5 @@
 import clsx from "clsx";
-
-const colors = [
-  "purple",
-  "blue",
-  "green",
-  "yellow",
-  "orange",
-  "red",
-  "brown",
-  "cyan",
-  "gold",
-  "olive",
-  "rose",
-];
+import { getUserInitials, getUserColor } from "./utils";
 
 export type AvatarProps = {
   fullName: string;
@@ -20,30 +7,13 @@ export type AvatarProps = {
   forceColor?: string;
 };
 
-// TODO: Add unit test
-const getInitials = (name: string) => {
-  const parts = name.split(" ");
-  if (parts.length > 2) {
-    // If there are more than 2 words, take only the first two initials
-    return parts
-      .slice(0, 2)
-      .map((n) => n[0])
-      .join("");
-  }
-  return parts.map((n) => n[0]).join("");
-};
-
-const getColor = (name: string) => {
-  return colors[name.charCodeAt(0) % colors.length];
-};
-
 export const UserAvatar = ({
   fullName,
   size = "medium",
   forceColor,
 }: AvatarProps) => {
-  const initials = getInitials(fullName);
-  const backgroundColor = getColor(fullName);
+  const initials = getUserInitials(fullName);
+  const backgroundColor = getUserColor(fullName);
   return (
     <div
       style={{ backgroundColor: forceColor ? forceColor : undefined }}